Multi genome project shows now reference stripes and operations including the "zero" genotypes can be implemented.
The TrackMaskFilter has been commented but not removed yet.
It has to be done soon! The "Advanced Filters" node is empty for now (that filter was the only one).

Better management of the VCF line with a new object: VCFLine.
Many lines and classes reformatted for better performance and more visibility.

Cache system for track graphics in use for VCF stripes. Can be easily applied for other tracks.
Can be enable/disable on the project Option dialog.


Former-commit-id: 864755928cc0e7ebcb8e935476d5c82b14703827